Move LazySplitSubsets to data/ as it was all alone in subset/
