+++ src/filter.cpp
@@ -8,8 +8,7 @@
 
 #include <dplyr/data/GroupedDataFrame.h>
 #include <dplyr/data/NaturalDataFrame.h>
-
-#include <dplyr/subset/LazySplitSubsets.h>
+#include <dplyr/data/LazySplitSubsets.h>
 
 #include <tools/bad.h>
 #include <tools/set_rownames.h>
